requirements.txt
coverage==5.0.3
Django==2.2.10
django-cors-headers==3.2.1
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0